extract MyPlayer as ExoPlayerUsingTextureView and move to library module
